# nsqd: remove channel messagePump