(Pushing up rebased branches I've had laying around locally for ages)

This removes a goroutine per channel and simplifies exit logic.