@mreiferson 2016-04-18
Also, w/r/t concurrency, because each _client_ is now selecting on mem/disk, a slow disk read won't block concurrent clients from being able to proceed reading from incoming messages on the mem channel.