@judwhite 2016-04-26
I'll have to sit down with this one to fully understand it. I planned on looking into the ID perf one later this week and doing some bench tests. So, no, nothing yet from me :smiley: