@ploxiln 2016-04-26
I'm no expert, but I like this change. Removes a goroutine and chan per nsq channel, a handoff per message, net lines of code, and seems to accomplish the same end result.