@mreiferson 2016-04-27
yep

```
benchmark                           old ns/op     new ns/op     delta
BenchmarkProtocolV2MultiSub1-4      5115          4364          -14.68%
BenchmarkProtocolV2MultiSub2-4      7466          6449          -13.62%
BenchmarkProtocolV2MultiSub4-4      13076         11632         -11.04%
BenchmarkProtocolV2MultiSub8-4      25147         22599         -10.13%
BenchmarkProtocolV2MultiSub16-4     55787         45889         -17.74%

benchmark                           old MB/s     new MB/s     speedup
BenchmarkProtocolV2MultiSub1-4      50.04        58.65        1.17x
BenchmarkProtocolV2MultiSub2-4      68.57        79.39        1.16x
BenchmarkProtocolV2MultiSub4-4      78.31        88.03        1.12x
BenchmarkProtocolV2MultiSub8-4      81.44        90.62        1.11x
BenchmarkProtocolV2MultiSub16-4     73.42        89.26        1.22x
```